+++ R/wms.R
@@ -15,6 +15,8 @@ wmsDependency <- function() {
 #' single-tile/untiled/nontiled layers, shared WMS sources, and
 #' GetFeatureInfo-powered identify.
 #' @inheritParams leaflet::addWMSTiles
+#' @param checkempty Should the returned HTML-content be checked for emptiness?
+#'   If the HTML-body is empty no popup is opened. Default is \code{FALSE}
 #' @param popupOptions List of popup options. See
 #'   \code{\link[leaflet]{popupOptions}}. Default is NULL.
 #' @inherit leaflet::addWMSTiles return
@@ -31,6 +33,7 @@ wmsDependency <- function() {
 #'   addWMS(baseUrl = "https://maps.dwd.de/geoserver/dwd/wms",
 #'          layers = "dwd:BRD_1km_winddaten_10m",
 #'       popupOptions = popupOptions(maxWidth = 600),
+#'       checkempty = TRUE,
 #'       options = WMSTileOptions(
 #'         transparent = TRUE,
 #'         format = "image/png",
@@ -40,6 +43,7 @@ addWMS <- function(map, baseUrl, layerId = NULL, group = NULL,
                    attribution = NULL,
                    layers = NULL,
                    popupOptions = NULL,
+                   checkempty = FALSE,
                    data = getMapData(map)) {
 
   if (is.null(layers)) {
@@ -47,6 +51,7 @@ addWMS <- function(map, baseUrl, layerId = NULL, group = NULL,
   }
   options$attribution <- attribution
   options$layers <- layers
+  options$checkempty <- checkempty
 
   map$dependencies <- c(map$dependencies, wmsDependency())