+++ inst/examples/wms_popup.R
@@ -15,6 +15,7 @@ server <- function(input, output, session) {
       setView(9, 50, 5) %>%
       addWMS(baseUrl = "https://maps.dwd.de/geoserver/dwd/wms", layers = "dwd:BRD_1km_winddaten_10m",
              popupOptions = popupOptions(maxWidth = 600),
+             checkempty = TRUE,
              options = WMSTileOptions(
                transparent = TRUE,
                format = "image/png",